+++ token/services/identity/x509/crypto/bccsp.go
@@ -9,8 +9,8 @@ package crypto
 import (
 	"encoding/hex"
 
+	"github.com/hyperledger-labs/fabric-token-sdk/token/services/identity/x509/crypto/csp"
 	"github.com/hyperledger-labs/fabric-token-sdk/token/services/identity/x509/crypto/pkcs11"
-	sw2 "github.com/hyperledger-labs/fabric-token-sdk/token/services/identity/x509/crypto/sw"
 	"github.com/hyperledger/fabric/bccsp"
 	"github.com/hyperledger/fabric/bccsp/sw"
 	"github.com/pkg/errors"
@@ -71,7 +71,7 @@ func GetDefaultBCCSP(keyStore bccsp.KeyStore) (bccsp.BCCSP, error) {
 	if keyStore == nil {
 		keyStore = sw.NewDummyKeyStore()
 	}
-	cryptoProvider, err := sw2.NewCSP(keyStore)
+	cryptoProvider, err := csp.NewCSP(keyStore)
 	if err != nil {
 		return nil, err
 	}
