+++ token/services/identity/x509/crypto/csp/ecdsakey.go
@@ -13,7 +13,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package sw
+package csp
 
 import (
 	"crypto/ecdsa"
