+++ token/services/identity/x509/crypto/csp/keyderiv.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package sw
+package csp
 
 import (
 	"crypto/ecdsa"
