+++ token/services/identity/x509/crypto/csp/keys.go
@@ -4,7 +4,7 @@ Copyright IBM Corp. All Rights Reserved.
 SPDX-License-Identifier: Apache-2.0
 */
 
-package sw
+package csp
 
 import (
 	"crypto/ecdsa"
